crypto: move merkle tree into the hash package
It's all about hashes, so it makes sense putting it there.
