+++ pkg/crypto/hash/merkle_tree_test.go
@@ -1,4 +1,4 @@
-package crypto
+package hash
 
 import (
 	"testing"
